# com.ptoceti.osgi.ihm/src/main/webapp/ihm/api-swagger.yaml
openapi: 3.0.1
info:
  title: Ptoceti Rest Api
  version: 1.0.0
paths:
...
  /configuration/conf/{pid}:
    get:
      tags:
        - configuration
      operationId: getConfiguration
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
    put:
      tags:
        - configuration
      operationId: updateConfiguration
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/ConfigurationWrapper'
      responses:
        default:
          description: default response
          content:
            '*/*': { }
    delete:
      tags:
        - configuration
      operationId: deleteConfiguration
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            '*/*': { }
...
  /devices/{pid}/configuration:
    get:
      tags:
        - devices
      operationId: getConfiguration_1
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
...
  /drivers/{pid}:
    get:
      tags:
        - driver
      operationId: getDriver
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/DriverWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/DriverWrapper'
  /drivers/{pid}/configuration:
    get:
      tags:
        - driver
      operationId: getConfiguration_2
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
...
  /things/{pid}:
    get:
      tags:
        - things
      operationId: getThing
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ThingWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/ThingWrapper'
  /things/{pid}/configuration:
    get:
      tags:
        - things
      operationId: getConfiguration_3
      parameters:
        - name: pid
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationWrapper'
...
  /wires/{id}:
    get:
      tags:
        - wires
      operationId: getWire
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: default response
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/WireWrapper'
            application/json:
              schema:
                $ref: '#/components/schemas/WireWrapper'
components:
  schemas:
...
    ConfigurationPropertyEntry:
      type: object
      properties:
        key:
          type: string
          xml:
            attribute: true
        value:
          type: string
          xml:
            attribute: true
    ConfigurationWrapper:
      type: object
      properties:
        bundleLocation:
          type: string
          xml:
            attribute: true
        factoryPid:
          type: string
          xml:
            attribute: true
        pid:
          type: string
          xml:
            attribute: true
        properties:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/ConfigurationPropertyEntry'
...
    DriverWrapper:
      type: object
      properties:
        deviceSerial:
          type: string
          xml:
            attribute: true
        driverId:
          type: string
          xml:
            attribute: true
        properties:
          type: object
          additionalProperties:
            type: string
...
    ThingWrapper:
      type: object
      properties:
        description:
          type: string
          xml:
            attribute: true
        driver:
          type: string
          xml:
            attribute: true
        firmwareVendor:
          type: string
          xml:
            attribute: true
        firmwareVersion:
          type: string
          xml:
            attribute: true
        hardwareVendor:
          type: string
          xml:
            attribute: true
        hardwareVersion:
          type: string
          xml:
            attribute: true
        model:
          type: string
          xml:
            attribute: true
        name:
          type: string
          xml:
            attribute: true
        properties:
          type: object
          additionalProperties:
            type: string
        serialNumber:
          type: string
          xml:
            attribute: true
        status:
          type: string
          xml:
            attribute: true
          enum:
            - STATUS_REMOVED
            - STATUS_OFFLINE
            - STATUS_ONLINE
            - STATUS_PROCESSING
            - STATUS_NOT_INITIALIZED
            - STATUS_NOT_CONFIGURED
        statusDetail:
          type: string
          xml:
            attribute: true
          enum:
            - STATUS_DETAIL_CONNECTING
            - STATUS_DETAIL_INITIALIZING
            - STATUS_DETAIL_REMOVING
            - STATUS_DETAIL_FIRMWARE_UPDATING
            - STATUS_DETAIL_CONFIGURATION_UNAPPLIED
            - STATUS_DETAIL_BROKEN
            - STATUS_DETAIL_COMMUNICATION_ERROR
            - STATUS_DETAIL_DATA_INSUFFICIENT
            - STATUS_DETAIL_INACCESSIBLE
            - STATUS_DETAIL_CONFIGURATION_ERROR
            - STATUS_DETAIL_DUTY_CYCLE
        types:
          type: array
          items:
            type: string
        uid:
          type: string
          xml:
            attribute: true
      xml:
        name: Thing
    WireWrapper:
      type: object
      properties:
        connected:
          type: boolean
          xml:
            attribute: true
        consumerPid:
          type: string
          xml:
            attribute: true
        lastValue:
          type: object
        pid:
          type: string
          xml:
            attribute: true
        producerPid:
          type: string
          xml:
            attribute: true
        scope:
          type: array
          items:
            type: string
        valid:
          type: boolean
          xml:
            attribute: true
      xml:
        name: Wire